Tweak Show instances to pad with zeros
Nanoseconds is to the right of the decimal, and so needs to be zero
padded if represenations like 532s.23456789ns are to make sense; needs
to be 532s.023456789ns. So pad to a width of nine characers. Obviously
Show instances are just for debugging, but it's nice to be uniform.

While we're at it, zero pad the Hour, Minute, and Second types' Show
instances to width two characters.