Add zero UTC offset (Z) support to ISO 8601 format
Previously, the documentation said it could parse time strings in the
ISO 8601 format, and gave two examples:

- 2014-04-05T17:25:04+00:00
- 2014-04-05T17:25:04Z

Unfortunately the support for the latter was not there yet, because the
`Format_TzHM_Colon` constructor of `TimeFormatElem` could not handle the
character "Z", the zero UTC offset zone designator in ISO 8601.

This commit leaves the `Format_TzHM_Colon` constructor as is for
backwards compatibility, and adds a new constructor
`Format_TzHM_Colon_Z` which first attempts to parse "Z" and falls back
to `Format_TzHM_Colon`. Then it changes the `TimeFormat` instance of
`ISO8601_DateAndTime` to give it the proper capability to parse the
second example.

As far as I can tell, this is a backwards compatible change.

(Hopefully) Solves #30.